requirements.txt
asgiref==3.8.1
Django==5.1.4
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.7.0
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.2
typing_extensions==4.12.2
